Add files via upload
This repository contains a Generative Adversarial Network (GAN) designed to generate images. The GAN model is trained on a dataset of single digit numbers (MNIST dataset) and leverages the power of both the generator and discriminator to create  images. Libraries used in this code are torch, numpy, torchvision, matplotlib and tqdm.